+++ src/components/molecules/NotificationMarkdown/NotificationMarkdown.tsx
@@ -10,9 +10,8 @@ import store from '@redux/store';
 
 import {TelemetryButtons} from '@molecules/NotificationMarkdown/TelemetryButtons';
 
-import {openUrlInExternalBrowser} from '@utils/shell';
-
 import {AlertType, ExtraContentType} from '@shared/models/alert';
+import {openUrlInExternalBrowser} from '@shared/utils/shell';
 
 import NotificationModalTitle from './NotificationModalTitle';
