+++ src/components/organisms/NewResourceWizard/NewResourceWizard.tsx
@@ -20,8 +20,6 @@ import {saveUnsavedResources} from '@redux/thunks/saveUnsavedResources';
 import {useFolderTreeSelectData} from '@hooks/useFolderTreeSelectData';
 import {useNamespaces} from '@hooks/useNamespaces';
 
-import {openNamespaceTopic, openUniqueObjectNameTopic} from '@utils/shell';
-
 import {getResourceKindHandler} from '@src/kindhandlers';
 
 import {ROOT_FILE_ENTRY} from '@shared/constants/fileEntry';
@@ -31,6 +29,7 @@ import {FileEntry} from '@shared/models/fileEntry';
 import {K8sResource} from '@shared/models/k8sResource';
 import {ResourceKindHandler} from '@shared/models/resourceKindHandler';
 import {NewResourceWizardInput} from '@shared/models/ui';
+import {openNamespaceTopic, openUniqueObjectNameTopic} from '@shared/utils/shell';
 
 import {FileCategoryLabel, FileNameLabel, SaveDestinationWrapper, StyledSelect} from './NewResourceWizard.styled';
