+++ src/components/organisms/StartProjectPane/RecentProjectsPage.tsx
@@ -11,14 +11,13 @@ import {
   toggleStartProjectPane,
 } from '@redux/reducers/ui';
 
-import {trackEvent} from '@utils/telemetry';
-
 import SelectFolder from '@assets/FromFolder.svg';
 import CreateFromGit from '@assets/FromGit.svg';
 import QuickClusterPreview from '@assets/QuickClusterPreview.svg';
 
 import {Project} from '@shared/models/config';
 import {activeProjectSelector} from '@shared/utils/selectors';
+import {trackEvent} from '@shared/utils/telemetry';
 
 import ActionCard from './ActionCard';
 import Guide from './Guide';
