+++ src/components/organisms_new/TutorialPage/TutorialCard.tsx
@@ -3,7 +3,6 @@ import {shell} from 'electron';
 import React, {useCallback} from 'react';
 
 import {Feature} from '@shared/models/features';
-import {trackEvent} from '@shared/utils/telemetry';
 
 import * as S from './TutorialCard.styled';
 
@@ -16,10 +15,6 @@ const TutorialCard: React.FC<Props> = ({feature}) => {
 
   const openLearnMore = useCallback(() => shell.openExternal(learnMoreUrl), [learnMoreUrl]);
 
-  const onCTAHandler = () => {
-    trackEvent('TUTORIAL_PAGE_OPENED', {id});
-  };
-
   return (
     <S.Card key={id}>
       <S.ElipseWrapper>
@@ -28,7 +23,7 @@ const TutorialCard: React.FC<Props> = ({feature}) => {
       <S.Name>{name}</S.Name>
       <S.Description>{description}</S.Description>
       <S.Span>
-        <S.Button onClick={onCTAHandler}>{callToAction}</S.Button>
+        <S.Button>{callToAction}</S.Button>
         <S.Link onClick={openLearnMore}>Learn more</S.Link>
       </S.Span>
     </S.Card>
