+++ src/kindhandlers/NodeHandler.ts
@@ -4,8 +4,8 @@ import log from 'loglevel';
 
 import navSectionNames from '@constants/navSectionNames';
 
-import {K8sResource} from '@models/k8sresource';
-import {ResourceKindHandler} from '@models/resourcekindhandler';
+import {K8sResource} from '@shared/models/k8sResource';
+import {ResourceKindHandler} from '@shared/models/resourceKindHandler';
 
 const NodeHandler: ResourceKindHandler = {
   kind: 'Node',
